feat(data_structures): implement `Send` and `Sync` for stack types (#17135)
Implement `Send` and `Sync` for `Stack<T>` and `NonEmptyStack<T>` where `T` is `Send` / `Sync`.

This also makes `SparseStack<T>` implement `Send` / `Sync` with same bounds.

`Vec` also implements `Send` / `Sync` in the same way, and the stack types are all essentially the same as `Vec`, just with a different internal implementations, so it makes sense to follow `Vec`.