docs(linter): List all rules in the config schema.
This will enable us to generate better validations in the future if we want to have rule configurations for native rules validated in the JSON schema.

It also enables better suggestions when users are configuring rules in their config files.

Unfortunately, this does mean an extra snapshot to update for every new rule added, but that shouldn'tbe *too* bad.