test(linter): Error when passed non-array config objects (#17198)
This has been a problem in various rules' tests, so I think we just fix it here and be done with it. You are only allowed to pass an array value for rule configs in the Tester, nothing else.

As an aside, I think the anchor-is-valid rule is basically broken, in that the config object it takes is _completely_ different from the one in [the original rule](https://github.com/jsx-eslint/eslint-plugin-jsx-a11y/blob/main/docs/rules/anchor-is-valid.md). We should maybe fix that.