fix(lsp): do not sent de-registrations for watched files on shutdown request (#17318)
TLDR: it is the client's fault to respect server registrations after shutdown

Some client (like helix editor) do not send an acknowledgment to the deregistration, because it already sent the shutdown request. (Some) notification are somehow possible. [The LSP Specs](https://microsoft.github.io/language-server-protocol/specifications/lsp/3.17/specification/#shutdown) are unclear about this behavior.
So to support fully one more editor, we expect that the deregistration of the file watchers are automatically handled by the client. Why should he still watch for them, when can not report to the server :)

closes https://github.com/oxc-project/oxc/issues/17190