perf(linter): inline `is_function_node` into run functions (#17373)
Inlining the `is_function_node` function allows the linter codegen to do the static analysis more easily. I suppose this is slightly negative towards less maintainability, but I don't expect we'll have more types of functions added to JS anytime soon. And the +1-2% improvement is hard to pass up for being so easy:

<img width="723" height="436" alt="image" src="https://github.com/user-attachments/assets/7612c36e-3006-4cea-a038-d0e9b027ee04" />