fix(oxfmt,oxlint): disable mimalloc for 32-bit Arm targets (#17473)
This makes the condition for enabling mimalloc the same across the
entire codebase. The condition for enabling mimalloc dependency was the
same across all the projects, but wasn't the same when actually enabling
it.
~Not sure if removing the miri condition is fine. But if we are
rewriting this, let's just do this as well and test it on CI~ (miri
condition added back after maintainer request)