refactor(semantic): consolidate diagnostics into single module (#17490)
Diagnostic functions for semantic were previously spread across 3x files, with `diagnostics.rs` only containing a single diagnostic. This PR refactors it to have all diagnostics in a single file. When diagnostics are needded, they are referenced via `diagnostic::my_diagnostic` to be clear and precise as to where the diagnostic is coming from.

I **think** this is better than how it was before - but if you disagree Boshen, feel free to close.