refactor(linter): update `Rule::from_configuration` to return a `Result` (#17479)
Built on top of #17478. 

Based on the changes I experimented with in #17199, and the suggestion
given by camc in that PR.

AI Disclosure: This was done almost entirely via find+replace and manual
updates to fix individual rules. I did use AI for the change in
0abc9814d3b9e44c59dcd042ef65f446a556d1a3 since I had trouble figuring
out a solution for that.

It may be easier to review this by going through it commit-by-commit.

Ran ecosystem CI on this branch and it looks like it works fine, no new
regressions
https://github.com/oxc-project/oxc-ecosystem-ci/actions/runs/20590908980

---------

Co-authored-by: Cameron Clark <cameron.clark@hey.com>