chore(linter): Add more test cases for import/extensions rule (#17536)
Some of these may be _technically_ duplicative of the existing test cases, but I think that's fine to confirm the behavior for such a complex rule. I did this as part of investigating #17501 and confirming the behavior is working as-expected in most cases.

There is one case I've left commented-out as the test fails, but it needs to be investigated as to whether that behavior matches the original ESLint rule or not.