perf(semantic): use SIMD for finding backslashes in `check_string_literal` (#17534)
Profiling shows that `check_string_literal` is called fairly often. We use a naive byte-by-byte check here for finding backslashes for octal escape sequences. This should be optimized to something wider than processing a single byte at a time, but it's worth ensuring this actually happens.

This PR changes it so that when the minimum string length is met, we use the SIMD-optimized `memchr` crate to find backslashes instead.

<img width="696" height="352" alt="image" src="https://github.com/user-attachments/assets/58c67f78-7b91-4ab1-aa45-68a175c38e6c" />